+++ plugins/hh-carnival/src/main/kotlin/ru/hh/android/plugin/services/CarnivalStartupActivity.kt
@@ -3,7 +3,7 @@ package ru.hh.android.plugin.services
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
-import ru.hh.android.plugin.config.PluginConfig
+import ru.hh.android.plugin.config.CarnivalPluginConfig
 import ru.hh.plugins.logger.HHLogger
 import ru.hh.plugins.logger.HHNotifications
 
@@ -17,11 +17,12 @@ class CarnivalStartupActivity : StartupActivity {
     }
 
     private fun setupLogger(project: Project) {
-        val config = PluginConfig.getInstance(project)
+        val config = CarnivalPluginConfig.getInstance(project)
         HHLogger.plant(project, "CarnivalLog", config.isDebugEnabled)
     }
 
     private fun setupNotifications(project: Project) {
         HHNotifications.plant(project, "Carnival")
     }
+
 }