+++ plugins/hh-garcon/src/main/kotlin/ru/hh/plugins/garcon/config/editor/GarconPluginSettings.kt
@@ -1,6 +1,7 @@
 package ru.hh.plugins.garcon.config.editor
 
 import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.openapi.components.service
@@ -10,6 +11,7 @@ import ru.hh.plugins.garcon.config.GarconPluginConfig
 import ru.hh.plugins.garcon.config.extensions.isNotFullyInitialized
 import ru.hh.plugins.utils.yaml.YamlUtils
 
+@Service(Service.Level.PROJECT)
 @State(
     name = "ru.hh.plugins.garcon.config.editor.GarconPluginSettings",
     storages = [Storage("garcon_plugin_settings.xml")]
@@ -23,7 +25,9 @@ class GarconPluginSettings : PersistentStateComponent<GarconPluginSettings> {
         fun getInstance(project: Project): GarconPluginSettings {
             return project.service<GarconPluginSettings>().let { settings ->
                 if (project.isDefault.not() && settings.config.isNotFullyInitialized()) {
-                    settings.tryLoadFromConfigFile(DEFAULT_PATH_TO_CONFIG_FILE)
+                    YamlUtils.tryLoadFromConfigFile<GarconPluginConfig>(DEFAULT_PATH_TO_CONFIG_FILE).onSuccess {
+                        settings.config = it
+                    }
                 }
 
                 settings
@@ -37,24 +41,11 @@ class GarconPluginSettings : PersistentStateComponent<GarconPluginSettings> {
 
     var config: GarconPluginConfig = GarconPluginConfig()
 
-    override fun getState(): GarconPluginSettings? {
+    override fun getState(): GarconPluginSettings {
         return this
     }
 
     override fun loadState(state: GarconPluginSettings) {
         XmlSerializerUtil.copyBean(state, this)
-    }
-
-    fun tryLoadFromConfigFile(configFilePath: String) {
-        YamlUtils.loadFromConfigFile<GarconPluginConfig>(
-            configFilePath = configFilePath,
-            onError = {
-                // todo
-            }
-        )?.let { configFromYaml ->
-            this.config = configFromYaml.copy(
-                configFilePath = configFilePath
-            )
-        }
     }
 }