+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/actions/SetupGeminioConfigAction.kt
@@ -5,18 +5,15 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.options.ShowSettingsUtil
 import ru.hh.plugins.geminio.config.editor.GeminioPluginSettingsSearchableConfigurable
 
-class SetupGeminioConfigAction : AnAction() {
-
-    init {
-        with(templatePresentation) {
-            text = "Setup Config"
-            isEnabledAndVisible = true
-        }
-    }
+class SetupGeminioConfigAction : AnAction(
+    /* text = */
+    "Setup Config"
+) {
 
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
         ShowSettingsUtil.getInstance()
             .editConfigurable(project, GeminioPluginSettingsSearchableConfigurable(project))
     }
+
 }