+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/config/editor/GeminioPluginSettings.kt
@@ -1,6 +1,7 @@
 package ru.hh.plugins.geminio.config.editor
 
 import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.openapi.components.service
@@ -8,8 +9,9 @@ import com.intellij.openapi.project.Project
 import com.intellij.util.xmlb.XmlSerializerUtil
 import ru.hh.plugins.geminio.config.GeminioPluginConfig
 import ru.hh.plugins.geminio.config.extensions.isNotFullyInitialized
-import ru.hh.plugins.geminio.config.extensions.tryLoadFromConfigFile
+import ru.hh.plugins.utils.yaml.YamlUtils
 
+@Service(Service.Level.PROJECT)
 @State(
     name = "ru.hh.plugins.geminio.config.editor.GeminioPluginConfig",
     storages = [Storage("geminio_plugin_settings.xml")]
@@ -23,10 +25,11 @@ class GeminioPluginSettings : PersistentStateComponent<GeminioPluginSettings> {
         fun getInstance(project: Project): GeminioPluginSettings {
             return project.service<GeminioPluginSettings>().let { settings ->
                 if (project.isDefault.not() && settings.config.isNotFullyInitialized()) {
-                    tryLoadFromConfigFile(DEFAULT_PATH_TO_CONFIG_FILE).onSuccess {
+                    YamlUtils.tryLoadFromConfigFile<GeminioPluginConfig>(DEFAULT_PATH_TO_CONFIG_FILE).onSuccess {
                         settings.config = it
                     }
                 }
+
                 settings
             }
         }
@@ -45,4 +48,5 @@ class GeminioPluginSettings : PersistentStateComponent<GeminioPluginSettings> {
     override fun loadState(state: GeminioPluginSettings) {
         XmlSerializerUtil.copyBean(state, this)
     }
+
 }