+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/config/extensions/GeminioPluginConfigExt.kt
@@ -1,10 +1,24 @@
 package ru.hh.plugins.geminio.config.extensions
 
 import ru.hh.plugins.geminio.config.GeminioPluginConfig
-import ru.hh.plugins.utils.yaml.YamlUtils
-import java.io.FileNotFoundException
+import ru.hh.plugins.geminio.config.editor.GeminioSettingsFormState
 
-fun GeminioPluginConfig.isNotFullyInitialized(): Boolean {
+internal fun GeminioPluginConfig.copyFromFormState(
+    state: GeminioSettingsFormState
+): GeminioPluginConfig {
+    return this.copy(
+        configFilePath = state.configFilePath.get(),
+        templatesRootDirPath = state.templatesRootDirPath.get(),
+        modulesTemplatesRootDirPath = state.modulesTemplatesRootDirPath.get(),
+        groupsNames = this.groupsNames.copy(
+            forNewGroup = state.nameForNewGroup.get(),
+            forNewModulesGroup = state.nameForNewModulesGroup.get(),
+        ),
+        isDebugEnabled = state.isDebugEnabled.get()
+    )
+}
+
+internal fun GeminioPluginConfig.isNotFullyInitialized(): Boolean {
     return configFilePath.isBlank() ||
         templatesRootDirPath.isBlank() ||
         modulesTemplatesRootDirPath.isBlank() ||
@@ -13,19 +27,4 @@ fun GeminioPluginConfig.isNotFullyInitialized(): Boolean {
 
 private fun GeminioPluginConfig.GroupsNames.isEmpty(): Boolean {
     return forNewGroup.isBlank() || forNewModulesGroup.isBlank()
-}
-
-internal fun tryLoadFromConfigFile(configFilePath: String): Result<GeminioPluginConfig> {
-    val configFromYaml = YamlUtils.loadFromConfigFile<GeminioPluginConfig>(
-        configFilePath = configFilePath,
-        onError = {
-            return Result.failure(it)
-        }
-    )?.copy(configFilePath = configFilePath)
-
-    return if (configFromYaml != null) {
-        Result.success(configFromYaml)
-    } else {
-        Result.failure(FileNotFoundException("File `$configFilePath` not found"))
-    }
 }