+++ shared/core/ui/src/main/kotlin/ru/hh/plugins/views/layouts/KotlinDestinationFolderComboBox.kt
@@ -3,7 +3,7 @@ package ru.hh.plugins.views.layouts
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiDirectory
 import com.intellij.refactoring.ui.PackageNameReferenceEditorCombo
-import com.intellij.ui.layout.Row
+import com.intellij.ui.dsl.builder.Row
 import org.jetbrains.kotlin.idea.refactoring.ui.KotlinDestinationFolderComboBox
 
 fun Row.kotlinDestinationFolderComboBox(