+++ shared/core/utils/src/main/kotlin/ru/hh/plugins/utils/yaml/YamlUtils.kt
@@ -4,9 +4,25 @@ import org.yaml.snakeyaml.LoaderOptions
 import org.yaml.snakeyaml.Yaml
 import org.yaml.snakeyaml.constructor.CustomClassLoaderConstructor
 import java.io.File
+import java.io.FileNotFoundException
 import java.io.FileReader
 
 object YamlUtils {
+
+    inline fun <reified T : YamlConfigModel> tryLoadFromConfigFile(configFilePath: String): Result<T> {
+        val configFromYaml = loadFromConfigFile<T>(
+            configFilePath = configFilePath,
+            onError = {
+                return Result.failure(it)
+            }
+        )?.setConfigFilePath<T>(configFilePath = configFilePath)
+
+        return if (configFromYaml != null) {
+            Result.success(configFromYaml)
+        } else {
+            Result.failure(FileNotFoundException("File `$configFilePath` not found"))
+        }
+    }
 
     inline fun <reified T> loadFromConfigFile(configFilePath: String, onError: (Throwable) -> Unit): T? {
         val configFile = File(configFilePath).takeIf { it.exists() } ?: return null