+++ shared/feature/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/mapping/widgets/predefined/PredefinedFeaturesParamsFactory.kt
@@ -97,4 +97,3 @@ internal fun PredefinedFeaturesSection.Companion.createSourceCodeFolderName(
         }
     )
 }
-